Add shirakumo/trial as a qlot dependency.
This will hopefully avoid having to add the first part of installation
instructions in the readme and will make scripting easier, that being
said because qlot does not let me specify to my knowledge the dynamic
size of memory for sbcl, I cannot execute the asdf:load-system command
necessary to launch my game. So I am considering removing qlot
entirely and writing script similar to Shinmera with a setup.lisp file
in the root directory.