Compiles without errors
The game now is in a compileable state. From here it is will be
possible to get working on the actually testing parts of the game and
seeing how daemons will actually function. This is quite late in the
game jam to be done this but it is what is.

A few standard changes have been made to the code base's formating and
comments, formatting is now specified to 100 lines. This is in
line (pun intended) with google's common lisp standards.

Comments now more frequently including warnings and todo's.

Threads and machines have been introduced, work on constraints is on
going.

Documentation has been added to daemons.

Some unnessary comments have removed.